refactor: merapihkan struktur kode
