--- src/app/_components/container.tsx
@@ -1,9 +1,0 @@
-type Props = {
-  children?: React.ReactNode;
-};
-
-const Container = ({ children }: Props) => {
-  return <div className="container mx-auto px-5">{children}</div>;
-};
-
-export default Container;
