--- src/app/_components/date-formatter.tsx
@@ -1,12 +1,0 @@
-import { parseISO, format } from "date-fns";
-
-type Props = {
-   dateString: string;
-};
-
-const DateFormatter = ({ dateString }: Props) => {
-   const date = parseISO(dateString);
-   return <time dateTime={dateString}>{format(date, "LLLL	d, yyyy")}</time>;
-};
-
-export default DateFormatter;
