--- src/app/_components/post-preview.tsx
@@ -1,41 +1,0 @@
-import { type Author } from "@/interfaces/author";
-import Link from "next/link";
-import Avatar from "./avatar";
-import CoverImage from "./cover-image";
-import DateFormatter from "./date-formatter";
-
-type Props = {
-   title: string;
-   coverImage: string;
-   date: string;
-   description: string;
-   author: Author;
-   slug: string;
-};
-
-export function PostPreview({
-   title,
-   coverImage,
-   date,
-   description,
-   author,
-   slug,
-}: Props) {
-   return (
-      <div>
-         <div className="mb-5">
-            <CoverImage slug={slug} title={title} src={coverImage} />
-         </div>
-         <h3 className="text-3xl mb-3 leading-snug">
-            <Link href={`/posts/${slug}`} className="hover:underline">
-               {title}
-            </Link>
-         </h3>
-         <div className="text-lg mb-4">
-            <DateFormatter dateString={date} />
-         </div>
-         <p className="text-lg leading-relaxed mb-4">{description}</p>
-         <Avatar name={author.name} picture={author.picture} />
-      </div>
-   );
-}
