--- src/app/_components/section-separator.tsx
@@ -1,3 +1,0 @@
-export function SectionSeparator() {
-  return <hr className="border-neutral-200 mt-28 mb-24" />;
-}
