+++ src/components/Button.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import React from "react";
+
 import { TextReveal } from "./ui/Typography";
 
 interface Button {
