+++ src/components/homepage/Partner.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { SectionHeading } from "../ui/Typography";
 import { SlideIn } from "../ui/Transition";
