+++ src/components/homepage/ScrollText.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { VelocityScroll } from "../ui/VelocityScroll";
 
 function ScrollText() {
